requirements.txt
streamlit
beautifulsoup4
requests
google-cloud-language
